+++ Core/Src/Merenja.h
@@ -17,7 +17,7 @@ typedef struct {
 /*
  * UID = Uredjaj ID
  */
-void setUID(uint32_t u0, uint32_t u1, uint32_t u2);
+void setUID(uint8_t u0, uint8_t u1, uint8_t u2);
 uint32_t getUID();
